Move create-testing-dwh to cli/commands/ and refactor
Split the ~240-line create_testing_dwh command (which was tripping
PLR0915 at 119 statements) out of cli/main.py into its own module
under cli/commands/.

main.py registers the command via a small register(app) hook on the
new module, so additional commands can be peeled off the same way without circular-import gymnastics.
